Updated application to use the entire screen as a button whose label is update with the current date and time when it is pushed.
I didn't understand why the button takes up the entire screen and why
the the XML layout we had previously defined no longer existed in the
app until I took a look at the pre-genereated source code that was
pasted over for this exercise. The setContentView is being feed the
button as the view for the app, instead of the information in
R.layout.activity_hello_android.
over for this