Tutorial with Spinner Extra Credit completed.
Reverted old extra credit changes, including returning RadioGroup
generation back to XML and removing surplus buttons that demonstrated
scrolling.

Tutorial 3: Added a list to store the saved restaurants in. Modified the
layout in the XML to fit the list on the screen and preserve the
integrity of the form from before.

Extra Credit Spinner: Changed list view to a spinner in both the XML and
the LunchList.java.