Tutorial 5 Extra Credit Different Layouts:
Changed the layout row for each different type of restaurant. Did not
do this by overriding the methods listed in the book, but instead used
an If/Else/Else statement to see which type was being sent to the
RestaurantHolder class and changed what the row value was being set
equal to.